services/recognition-service/requirements.txt
pyzbar
opencv-python
numpy
flask
flask_restful
flask_cors
pillow